imbalanced_data_sampler_comparison/requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.62.1
imbalanced-learn==0.14.1
japanize-matplotlib==1.1.3
joblib==1.5.3
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.2.6
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tzdata==2025.3
